Quantity: decimal → long

+++ VirtoCommerce.Storefront.Model/Contracts/DerivativeContractItem.cs
@@ -10,10 +10,10 @@ namespace VirtoCommerce.Storefront.Model.Contracts
 
         public string ProductId { get; set; }
 
-        public decimal ContractSize { get; set; }
+        public long ContractSize { get; set; }
 
-        public decimal PurchasedQuantity { get; set; }
+        public long PurchasedQuantity { get; set; }
 
-        public decimal RemainingQuantity { get; set; }
+        public long RemainingQuantity { get; set; }
     }
 }
